[SOIN] Extrais les questions
